Fix bugs in Fill, handle, stream

+++ samgraph/common/operation.h
@@ -7,7 +7,7 @@ namespace common {
 extern "C" {
 
 void samgraph_init(const char*path, int sample_device, int train_device,
-                   int batch_size, int *fanout, int num_fanout, int num_epoch);
+                   size_t batch_size, int *fanout, size_t num_fanout, int num_epoch);
 
 void samgraph_start();
